Switch from binary idpConfig.obj to XML idpConfig.xml
When adapting IdentityProviderDemo for use in another solution, the IdP is commonly renamed. However, because idpConfig.obj ships with the IdP and is binary serialized, idpConfig.obj includes the original namespace of FileConfig. After renaming, IdP fails on startup because namespace + type to deserialize into cannot be located.

With XML serialization, only type name is included in idpConfig.xml, not its namespace, making it rename resilient.